requirements.txt
transformers 
datasets 
evaluate 
accelerate 
pillow 
timm